Refactor logging and update `size` column type.
Commented out excessive logging for cleaner code and improved debugging clarity. Updated `size` column in `virtual_files` table to `UNSIGNED BIGINT` to support larger values. Added new dependency `predis/predis` to `composer.json`.

+++ database/migrations/2024_11_15_185656_create_virtual_files_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
             $table->string('extension');
             $table->string('minetypes');
             $table->string('disk');
-            $table->integer('size');
+            $table->unsignedBigInteger('size');
             $table->integer('expired_at');
             $table->timestamps();
             $table->engine('InnoDB');
